test vehicle commands and fixed bugs in it.
I guess there is a bug in `GetStops` command in sumo itself, cause it returns an TraciCompoudObject, which says it has 9 inner object, but in fact it has 16*numofstops objects. In python source code, sumo offfical wrote an callback just for this command to decode a TraciCompoudObject. I hope this will be fixed in future.